+++ tests/testthat/test-plumber.R
@@ -173,51 +173,7 @@ test_that("mounts can be read correctly", {
   expect_s3_class(pr$mounts[["/mysubpath/"]], "plumber")
 })
 
-test_that("prints correctly", {
-  testthat::skip_on_cran()
-  testthat::skip_on_os("windows") # has issues comparing text values
 
-  pr <- plumber$new()
-  pr$handle("GET", "/nested/path/here", function(){})
-  pr$handle("POST", "/nested/path/here", function(){})
-
-  pr2 <- plumber$new()
-  pr2$handle("POST", "/something", function(){})
-  pr2$handle("GET", "/", function(){})
-  pr$mount("/mysubpath", pr2)
-
-  stat <- PlumberStatic$new(".")
-  pr$mount("/static", stat)
-
-  printed <- capture.output(print(pr))
-
-  regexps <- c(
-    "Plumber router with 2 endpoints, 4 filters, and 2 sub-routers",
-    "Call run() on this object",
-    "├──[queryString]",
-    "├──[postBody]",
-    "├──[cookieParser]",
-    "├──[sharedSecret]",
-    "├──/nested",
-    "│  ├──/path",
-    "│  │  └──/here (GET, POST)",
-    "├──/mysubpath",
-    "│  │ # Plumber router with 2 endpoints, 4 filters, and 0 sub-routers.",
-    "│  ├──[queryString]",
-    "│  ├──[postBody]",
-    "│  ├──[cookieParser]",
-    "│  ├──[sharedSecret]",
-    "│  ├──/something (POST)",
-    "│  └──/ (GET)",
-    "├──/static",
-    "│  │ # Plumber static router serving from directory: ."
-  )
-
-  for (i in 1:length(regexps)){
-    expect_match(printed[i], regexps[i], info=paste0("on line ", i), fixed = TRUE)
-  }
-
-})
 
 test_that("mounts work", {
   pr <- plumber$new()